Read networks config from default home config file
It is expected that configuration for networks can be available under
`~/.hardhat/networks.json` which is a default path. Additionally to this
path user can also configure `localNetworksConfig` path in the project
configuration.

The code was refactored in preparation to change the order of how
configs are resolved and overwritten.

With this change the order remains the same:
1. Project network specific configuration
2. Local network specific configuration
3. Local default network configuration (`~/.hardhat/networks.json`)